requirements.txt
#brother_ql
bottle>=0.12.25
jinja2
pycups
pylibdmtx[scripts]
pillow>=10.3.0
requests